Refactor backup-restore script to handle old and new file backup process. Centralised config viewing module

+++ scripts/backup-restore.sh
@@ -3,18 +3,24 @@
 TEMP_FILE="$ENV_FILE.tmp"
 BACKUP_FILE="$ENV_FILE.backup"
 
+# TODO temporary to migrate old config over to new
+if grep -q "^#------------------------------------------------------------------------$" "$ENV_FILE"; then
+    IS_OLD_CONFIG=true
+else
+    IS_OLD_CONFIG=false
+fi
+
 display_banner() {
     clear
     echo "Backup & Restore Config Manager"
-    echo "${GREEN}----------------------------------------${NC}"
-    echo
+    echo "${GREEN}----------------------------------------${NC}\n"
 }
 
 backup_config() {
     if [ -f "$BACKUP_FILE" ]; then
         while true; do
             display_banner
-            options="(1-2)"
+            options="(1-3)"
 
             echo "Backup file ${RED}$BACKUP_FILE${NC} already exists.\n"
             echo "Choose an option:\n"
@@ -28,17 +34,22 @@ backup_config() {
             case "$option" in
                 1)
                     display_banner
-                    echo "Content of current backup file ${RED}$BACKUP_FILE${NC}.\n"
-                    echo "${YELLOW}---------[ START OF BACKUP ]---------\n${BLUE}"
-                    cat "$BACKUP_FILE"
-                    echo "${YELLOW}\n----------[ END OF BACKUP ]----------${NC}"
+                    $VIEW_CONFIG $BACKUP_FILE "BACKUP"
                     ;;
                 2)
                     display_banner
                     echo "Content of current in-use configuration.\n"
-                    echo "${YELLOW}---------[ START OF CONFIG ]---------\n${BLUE}"
-                    tail -n +14 $ENV_FILE
-                    echo "${YELLOW}\n----------[ END OF CONFIG ]----------${NC}"
+
+                    # TODO - Remove in future updates
+                    if [ "$IS_OLD_CONFIG" = true ]; then
+                        echo "${YELLOW}---------[ START OF CONFIG ]---------\n${BLUE}"
+                        tail -n +14 $ENV_FILE
+                        echo "${YELLOW}\n----------[ END OF CONFIG ]----------${NC}"
+
+                        printf "\nPress Enter to continue..."; read input
+                    else
+                        $VIEW_CONFIG
+                    fi
                     ;;
                 3)
                     display_banner
@@ -52,32 +63,37 @@ backup_config() {
                 *)
                     echo
                     echo "Invalid option. Please select a valid option $options."
+                    printf "\nPress Enter to continue..."; read input
                     ;;
             esac
-            printf "\nPress Enter to continue..."; read input
+            # printf "\nPress Enter to continue..."; read input
         done
     fi
 
     BACKUP_FLAG=0
     : > "$BACKUP_FILE" # Clear previous backup file or create new
 
-    while IFS= read -r line; do
-        # Check if end of system config
-        if [ "$line" = "#------------------------------------------------------------------------" ]; then
-            BACKUP_FLAG=1
-            continue
-        fi
+    if [ "$IS_OLD_CONFIG" = true ]; then
+        while IFS= read -r line; do
+            # Check if end of system config
+            if [ "$line" = "#------------------------------------------------------------------------" ]; then
+                BACKUP_FLAG=1
+                continue
+            fi
 
-        # Skip the frst line after system config
-        if [ "$BACKUP_FLAG" -eq 1 ] && [ -z "$line" ]; then
-            BACKUP_FLAG=2
-            continue
-        fi
+            # Skip the frst line after system config
+            if [ "$BACKUP_FLAG" -eq 1 ] && [ -z "$line" ]; then
+                BACKUP_FLAG=2
+                continue
+            fi
 
-        if [ "$BACKUP_FLAG" -eq 2 ]; then
-            echo "$line" >> "$BACKUP_FILE"
-        fi
-    done < "$ENV_FILE"
+            if [ "$BACKUP_FLAG" -eq 2 ]; then
+                echo "$line" >> "$BACKUP_FILE"
+            fi
+        done < "$ENV_FILE"
+    else
+        cp -f "$ENV_FILE" "$BACKUP_FILE"
+    fi
 
     echo "Backup completed. Content has been saved to ${RED}$BACKUP_FILE${NC}."
 }
@@ -88,19 +104,23 @@ restore_config() {
         return
     fi
 
-    # Get system config
-    awk '
-    /#------------------------------------------------------------------------/ {
-        print
-        print ""
-        exit
-    }
-    {print}
-    ' "$ENV_FILE" > "$TEMP_FILE"
+    if [ "$IS_OLD_CONFIG" = true ]; then
+        # Get system config
+        awk '
+        /#------------------------------------------------------------------------/ {
+            print
+            print ""
+            exit
+        }
+        {print}
+        ' "$ENV_FILE" > "$TEMP_FILE"
 
-    echo "$(cat $BACKUP_FILE)" >> "$TEMP_FILE"
-    mv -f "$TEMP_FILE" "$ENV_FILE"
-    rm -f "$BACKUP_FILE"
+        echo "$(cat $BACKUP_FILE)" >> "$TEMP_FILE"
+        mv -f "$TEMP_FILE" "$ENV_FILE"
+        rm -f "$BACKUP_FILE"
+    else
+        mv -f $BACKUP_FILE $ENV_FILE
+    fi
     echo "Restore completed. Content has been restored from ${RED}$BACKUP_FILE${NC}."
 }
